Separation of logic in Parser
