+++ src/main/java/kroppeb/stareval/parser/ParserOptions.java
@@ -10,18 +10,20 @@ import java.util.Map;
  * --
  */
 public class ParserOptions {
-	public ParserOptions(Char2ObjectMap<OpResolver<UnaryOp>> unaryOpResolvers, Char2ObjectMap<OpResolver<BinaryOp>> binaryOpResolvers) {
+	public ParserOptions(Char2ObjectMap<OpResolver<UnaryOp>> unaryOpResolvers, Char2ObjectMap<OpResolver<BinaryOp>> binaryOpResolvers, ParserParts parserParts) {
 		this.unaryOpResolvers = unaryOpResolvers;
 		this.binaryOpResolvers = binaryOpResolvers;
+		this.parserParts = parserParts;
 	}
 	
 	final Char2ObjectMap<OpResolver<UnaryOp>> unaryOpResolvers;
 	final Char2ObjectMap<OpResolver<BinaryOp>> binaryOpResolvers;
-	
+	final ParserParts parserParts;
 	
 	public static class Builder {
-		final Char2ObjectMap<Map<String, UnaryOp>> unaryOpResolvers = new Char2ObjectOpenHashMap<>();
-		final Char2ObjectMap<Map<String, BinaryOp>> binaryOpResolvers = new Char2ObjectOpenHashMap<>();
+		private final Char2ObjectMap<Map<String, UnaryOp>> unaryOpResolvers = new Char2ObjectOpenHashMap<>();
+		private final Char2ObjectMap<Map<String, BinaryOp>> binaryOpResolvers = new Char2ObjectOpenHashMap<>();
+		private ParserParts parserParts = null;
 		
 		public void addUnaryOp(String s, UnaryOp op) {
 			Map<String, UnaryOp> mp = this.unaryOpResolvers
@@ -35,6 +37,10 @@ public class ParserOptions {
 					.computeIfAbsent(s.charAt(0), (c) -> new Object2ObjectOpenHashMap<>());
 			BinaryOp previous = mp.put(s.substring(1), op);
 			assert previous == null;
+		}
+		
+		public void setParserParts(ParserParts parserParts){
+			this.parserParts = parserParts;
 		}
 		
 		private static <T extends Op> Char2ObjectMap<OpResolver<T>> convertOp(Char2ObjectMap<Map<String, T>> ops) {
@@ -81,8 +87,50 @@ public class ParserOptions {
 		public ParserOptions build() {
 			return new ParserOptions(
 					convertOp(this.unaryOpResolvers),
-					convertOp(this.binaryOpResolvers)
-			);
+					convertOp(this.binaryOpResolvers),
+					this.parserParts==null? new ParserParts() {}:this.parserParts);
+		}
+	}
+	
+	public interface ParserParts {
+		static boolean isNumber(final char c) {
+			return c >= '0' && c <= '9';
+		}
+		
+		static boolean isLowerCaseLetter(final char c) {
+			return c >= 'a' && c <= 'z';
+		}
+		
+		static boolean isUpperCaseLetter(final char c) {
+			return c >= 'A' && c <= 'Z';
+		}
+		
+		static boolean isLetter(final char c) {
+			return isLowerCaseLetter(c) || isUpperCaseLetter(c);
+		}
+		
+		default boolean isIdStart(final char c) {
+			return isLetter(c) || c == '_';
+		}
+		
+		default boolean isIdPart(final char c) {
+			return isIdStart(c) || isNumber(c);
+		}
+		
+		default boolean isNumberStart(final char c) {
+			return isNumber(c) || c == '.';
+		}
+		
+		default boolean isNumberPart(final char c) {
+			return isNumberStart(c) || isLetter(c);
+		}
+		
+		default boolean isAccessStart(final char c) {
+			return isIdStart(c) || isNumber(c);
+		}
+		
+		default boolean isAccessPart(final char c) {
+			return isAccessStart(c);
 		}
 	}
 }
